cls/requirements.txt
absl-py==1.0.0
adamp==0.3.0
albumentations==1.1.0
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
efficientnet-pytorch==0.7.1
einops==0.4.1
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
idna==3.3
imageio==2.19.1
importlib-metadata==4.11.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
loguru==0.6.0
Markdown==3.3.7
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munch==2.5.0
networkx==2.8
nibabel==3.2.2
oauthlib==3.2.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow
pretrainedmodels==0.7.4
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.64.0
typing_extensions 
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
gdown
